requirements.txt
streamlit
supabase
pandas
numpy
scikit-learn
plotly
python-dotenv